fixing interactive content
