+++ textContent/AllComponents.mdx
@@ -5,7 +5,7 @@ section: components
 sortValue: 1
 ---
 
-import { componentsData } from './components-data.jsx'
+import { componentsData } from './components-data.js'
 
 <SectionGallery
   section="components"
